// src/core/enums.ts
export enum Langs {
  EN = "EN",
  VI = "VI",
}

export enum PaymentMethodEnum {
  ONLINE = "ONLINE",
  ATM = "ATM",
  MOMO = "MOMO",
  SAMSUNG = "SAMSUNG",
  APPLE = "APPLE",
}

export enum InputTypeEnum {
  TEXT = "text",
  DATE = "date",
}

export enum TypeInsuranceEnum {
  LIABILITY = "LIABILITY",
  COMPREHENSIVE = "COMPREHENSIVE",
  HEALTH = "HEALTH",
}

export enum GtagFlowTypeEnum {
  MTPL = "MTPL",
  MPD = "MPD",
  HEALTH = "HEALTH",
}